fix(ci): split quota view models
